refactor: extract both inference sessions into a dedicated file (rationale: otherwise, their original files become hard to read)
